feat: add enable and upgrade support for theme list
