+++ src/modules/interface/themes/components/preview/ThemePreviewListItem.vue
